+++ packages/pandas-gbq/pandas_gbq/load.py
@@ -4,6 +4,7 @@ import io
 
 from google.cloud import bigquery
 
+from pandas_gbq.features import FEATURES
 import pandas_gbq.schema
 
 
@@ -30,10 +31,10 @@ def encode_chunk(dataframe):
     return io.BytesIO(body)
 
 
-def encode_chunks(dataframe, chunksize=None):
+def split_dataframe(dataframe, chunksize=None):
     dataframe = dataframe.reset_index(drop=True)
     if chunksize is None:
-        yield 0, encode_chunk(dataframe)
+        yield 0, dataframe
         return
 
     remaining_rows = len(dataframe)
@@ -41,10 +42,10 @@ def encode_chunks(dataframe, chunksize=None):
     start_index = 0
     while start_index < total_rows:
         end_index = start_index + chunksize
-        chunk_buffer = encode_chunk(dataframe[start_index:end_index])
+        chunk = dataframe[start_index:end_index]
         start_index += chunksize
         remaining_rows = max(0, remaining_rows - chunksize)
-        yield remaining_rows, chunk_buffer
+        yield remaining_rows, chunk
 
 
 def load_chunks(
@@ -60,24 +61,35 @@ def load_chunks(
     job_config.source_format = "CSV"
     job_config.allow_quoted_newlines = True
 
-    if schema is None:
+    # Explicit schema? Use that!
+    if schema is not None:
+        schema = pandas_gbq.schema.remove_policy_tags(schema)
+        job_config.schema = pandas_gbq.schema.to_google_cloud_bigquery(schema)
+    # If not, let BigQuery determine schema unless we are encoding the CSV files ourselves.
+    elif not FEATURES.bigquery_has_from_dataframe_with_csv:
         schema = pandas_gbq.schema.generate_bq_schema(dataframe)
+        schema = pandas_gbq.schema.remove_policy_tags(schema)
+        job_config.schema = pandas_gbq.schema.to_google_cloud_bigquery(schema)
 
-    schema = pandas_gbq.schema.add_default_nullable_mode(schema)
+    chunks = split_dataframe(dataframe, chunksize=chunksize)
+    for remaining_rows, chunk in chunks:
+        yield remaining_rows
 
-    job_config.schema = [
-        bigquery.SchemaField.from_api_repr(field) for field in schema["fields"]
-    ]
-
-    chunks = encode_chunks(dataframe, chunksize=chunksize)
-    for remaining_rows, chunk_buffer in chunks:
-        try:
-            yield remaining_rows
-            client.load_table_from_file(
-                chunk_buffer,
+        if FEATURES.bigquery_has_from_dataframe_with_csv:
+            client.load_table_from_dataframe(
+                chunk,
                 destination_table_ref,
                 job_config=job_config,
                 location=location,
             ).result()
-        finally:
-            chunk_buffer.close()
+        else:
+            try:
+                chunk_buffer = encode_chunk(chunk)
+                client.load_table_from_file(
+                    chunk_buffer,
+                    destination_table_ref,
+                    job_config=job_config,
+                    location=location,
+                ).result()
+            finally:
+                chunk_buffer.close()